pyproject.toml
[project]
name = "data-engineering-pipeline"
version = "0.1.0"
description = "Intermediate data engineering workshop: Airflow + Spark + Postgres + MinIO + Great Expectations + Python FastAPI backend. A learnwithparam.com workshop."
authors = [{ name = "Param Harrison", email = "param@learnwithparam.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "minio>=7.2.0",
    "structlog>=24.1.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
